requirements.txt
playwright==1.51.0
pytest==8.3.5
pytest-html==4.1.1
pytest-rerunfailures==15.0
python-dotenv==1.1.0